requirements.txt
imgaug==0.4.0
matplotlib==3.4.2
numpy==1.19.5
opencv_python==4.5.3.56
pandas==1.2.4
Pillow==9.0.1
pycocotools==2.0.4
scikit_learn==1.0.2
scipy==1.6.3
seaborn==0.11.1
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3